Extract applicationId and deploymentId to new subproject
