+++ contentgrid-spring-boot-autoconfigure/build.gradle
@@ -24,6 +24,7 @@ dependencies {
     implementation platform(project(':contentgrid-spring-boot-platform'))
     implementation 'org.springframework.boot:spring-boot-autoconfigure'
 
+    compileOnly project(':contentgrid-spring-common')
     compileOnly project(':contentgrid-spring-boot-actuators')
     compileOnly project(':contentgrid-spring-integration-events')
     compileOnly project(':contentgrid-spring-data-rest')
