+++ contentgrid-spring-boot-autoconfigure/src/main/java/com/contentgrid/spring/boot/autoconfigure/audit/ContentGridAuditMessagingAutoConfiguration.java
@@ -1,9 +1,11 @@
 package com.contentgrid.spring.boot.autoconfigure.audit;
 
+import com.contentgrid.spring.common.ContentGridApplicationPropertiesConfiguration;
 import com.contentgrid.spring.audit.handler.messaging.AuditEventMessageConverter;
 import com.contentgrid.spring.audit.handler.messaging.AuditEventToCloudEventMessageConverter;
 import com.contentgrid.spring.audit.handler.messaging.Jackson2AuditMessagingModule;
 import com.contentgrid.spring.audit.handler.messaging.MessageSendingAuditHandler;
+import com.contentgrid.spring.common.ContentGridApplicationProperties;
 import com.contentgrid.spring.boot.autoconfigure.audit.ContentGridAuditMessagingAutoConfiguration.ContentGridAuditMessagingProperties;
 import com.contentgrid.spring.boot.autoconfigure.messaging.ContentGridMessaging;
 import com.contentgrid.spring.boot.autoconfigure.messaging.ContentGridMessagingAutoConfiguration;
@@ -11,6 +13,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.cloudevents.CloudEvent;
 import io.cloudevents.spring.messaging.CloudEventMessageConverter;
 import java.net.URI;
+import java.util.Properties;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
@@ -20,23 +23,31 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
 import org.springframework.core.annotation.Order;
-import org.springframework.expression.ExpressionParser;
 import org.springframework.messaging.converter.MessageConverter;
 import org.springframework.messaging.core.MessageSendingOperations;
+import org.springframework.util.PropertyPlaceholderHelper;
+import org.springframework.util.SystemPropertyUtils;
 
 @AutoConfiguration(
         after = ContentGridMessagingAutoConfiguration.class,
         before = ContentGridAuditLoggingAutoConfiguration.class
 )
-@ConditionalOnClass({MessageSendingOperations.class, MessageSendingAuditHandler.class})
+@ConditionalOnClass({MessageSendingOperations.class, MessageSendingAuditHandler.class, ContentGridApplicationProperties.class})
 @ConditionalOnBean(value = MessageSendingOperations.class, annotation = ContentGridMessaging.class)
 @EnableConfigurationProperties(ContentGridAuditMessagingProperties.class)
 @ConditionalOnProperty(prefix = ContentGridAuditMessagingAutoConfiguration.CONTENTGRID_AUDIT_MESSAGING, name = "enabled", matchIfMissing = true)
+@Import(ContentGridApplicationPropertiesConfiguration.class)
 @Slf4j
 public class ContentGridAuditMessagingAutoConfiguration {
 
     public static final String CONTENTGRID_AUDIT_MESSAGING = "contentgrid.audit.messaging";
+
+    private static final PropertyPlaceholderHelper PROPERTY_PLACEHOLDER_HELPER = new PropertyPlaceholderHelper(
+            SystemPropertyUtils.PLACEHOLDER_PREFIX,
+            SystemPropertyUtils.PLACEHOLDER_SUFFIX
+    );
 
     @Bean
     @ConditionalOnProperty(prefix = CONTENTGRID_AUDIT_MESSAGING, name = "destination")
@@ -61,13 +72,22 @@ public class ContentGridAuditMessagingAutoConfiguration {
     MessageConverter auditEventToCloudEventMessageConverter(
             ObjectMapper objectMapper,
             ContentGridAuditMessagingProperties auditProperties,
-            ExpressionParser expressionParser
+            ContentGridApplicationProperties applicationProperties
     ) {
-        var parsedSource = expressionParser.parseExpression(auditProperties.getSource()).getValue(URI.class);
+        String parsedSource;
+        if (applicationProperties.getSystem().getApplicationId() != null
+                && applicationProperties.getSystem().getDeploymentId() != null) {
+            var props = new Properties();
+            props.put("applicationId", applicationProperties.getSystem().getApplicationId());
+            props.put("deploymentId", applicationProperties.getSystem().getDeploymentId());
+            parsedSource = PROPERTY_PLACEHOLDER_HELPER.replacePlaceholders(auditProperties.getSource(), props);
+        } else {
+            parsedSource = auditProperties.getSource();
+        }
         return new AuditEventToCloudEventMessageConverter(
                 new CloudEventMessageConverter(),
                 objectMapper::writeValueAsBytes,
-                parsedSource != null ? parsedSource : URI.create(auditProperties.getSource())
+                URI.create(parsedSource)
         );
     }
