+++ contentgrid-spring-common/build.gradle
@@ -7,9 +7,11 @@ configurations {
     compileOnly {
         extendsFrom(annotationProcessor)
     }
-
     testAnnotationProcessor {
         extendsFrom(annotationProcessor)
+    }
+    testImplementation {
+        extendsFrom(compileOnly)
     }
 }
 
@@ -18,18 +20,12 @@ dependencies {
     annotationProcessor 'org.projectlombok:lombok'
 
     implementation platform(project(':contentgrid-spring-boot-platform'))
-    implementation 'org.springframework.boot:spring-boot-autoconfigure'
+    implementation 'org.springframework.boot:spring-boot'
 
-    compileOnly project(':contentgrid-spring-common')
-
-    api 'org.springframework.boot:spring-boot-actuator'
-
+    testImplementation platform('org.junit:junit-bom:5.9.1')
     testImplementation 'org.junit.jupiter:junit-jupiter'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-
-    testCompileOnly 'org.projectlombok:lombok'
 }
 
-tasks.named('test') {
+test {
     useJUnitPlatform()
 }
