+++ contentgrid-spring-common/gradle.properties
@@ -1,0 +1,1 @@
+description=Shared dependencies for ContentGrid-Spring modules
