+++ settings.gradle
@@ -30,3 +30,4 @@ include 'contentgrid-spring-integration-events'
 include 'contentgrid-spring-querydsl'
 include 'contentgrid-spring-swagger-ui'
 include 'contentgrid-spring-audit-logging'
+include 'contentgrid-spring-common'
